requirements.txt
astroid==2.4.2
asyncio==3.4.3
asyncpg==0.21.0
asyncpgsa==0.26.3
attrs==20.3.0
autopep8==1.5.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
Cython==0.29.21
dnspython==2.0.0
idna==2.10
ipwhois==1.2.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pymongo==3.11.2
pyOpenSSL==20.0.0
pyTelegramBotAPI==3.7.4
requests==2.25.0
schedule==0.6.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
TelegramBotAPI==0.3.2
toml==0.10.2
urllib3==1.26.2
vedis==0.7.1
wrapt==1.12.1
